Edits focused on improving clarity of message.
These edits focused primarily on making word choice more concrete and improving the learning objectives of this lesson. There were a number of instances where very vague pronouns like "it" were being used in place of more concrete nouns, and this can be really confusing. It's better to use specific but accessible nouns. I also fixed a couple of typos and other inconsistencies in here. The biggest thing, though, was changing the learning objectives--here, I tried to match the learning objectives with the exercises being given and make their connection to Bloom's taxonomy more obvious.