Convert directly from Fahrenheit to kelvin
The Fahrenheit to Kelvin formula is based around converting the temperature first to Celsius, then to Kelvin. I've changed this to avoid the intermediate Celsius step (which could be confusing), and added a reference for the formula (for those who want to give a taste of looking for answers on the Web).