Update 04-cond.Rmd
I reordered the introduction of logicals and conditionals such that logicals and logical comparisons are introduced before their use in conditional statements. Additionally I changed instances in that section where the word true is used to 'TRUE' and the same with false. This should reinforce the status to TRUE and FALSE as distinct objects with a specific syntax in R to learners