Update 09-supp-intro-rstudio.md
I have a comment on the Introduction to RStudio episode of the Programming with R Software Carpentry Lesson. This episode provides a detailed overview of working in R Studio. However, no visuals are provided. Would it be possible to add either a screenshot of RStudio or a depiction of the RStudio working environment? For attendees of a Software Carpentry Session, the instructor can provide an overview during the lesson. However, for anyone who may be going through a lesson on their own or referring to the lesson as a reminder, there are no visuals on the episode page to help the learner follow along. The text does provide information on where each major component of RStudio exists in the working environment, i.e. "console window (in RStudio, the bottom left panel)." However, adding visual aids could be helpful for visual learners. See suggestions in the text for where visual aids could be included.